fix : gpt 모델 4.1로 변경, 전반적인 디자인 수정, 댓글 로직 리팩토링, mbti 질문 20개로 늘림.

+++ src/main/java/com/example/setting/dto/CommentDTO.java
@@ -15,7 +15,7 @@ public class CommentDTO {
     private String memberMbti;
     private LocalDateTime timestamp;
     private String memberNickname;
-    private int likes;
+    private Long likes;
     private String likesNickname;
     public CommentDTO(Comment comment){
         this.id = comment.getId();
